+++ .github/workflows/bundler-integrations/src/initAll.mjs
@@ -1,0 +1,4 @@
+// An example of importing and initialising allcomponents via initAll
+import { initAll } from 'govuk-frontend'
+
+initAll()